add admin_user op

+++ francisBlog/src/com/francis/blog/serviceImpl/UserManagerImpl.java
@@ -1,40 +1,65 @@
 package com.francis.blog.serviceImpl;
 
 import java.util.List;
+import java.util.Map;
 
 import javax.annotation.Resource;
 
 import org.springframework.stereotype.Component;
 
+import com.francis.blog.dao.ArticleDao;
+import com.francis.blog.dao.CommentsDao;
 import com.francis.blog.dao.UserDao;
+import com.francis.blog.pojo.Article;
+import com.francis.blog.pojo.Comments;
 import com.francis.blog.pojo.User;
 import com.francis.blog.service.UserManager;
 
 @Component("userManager")
 public class UserManagerImpl implements UserManager{
 	private UserDao userDao;
+	private CommentsDao commentsDao;
+	private ArticleDao articleDao;
+	
+	@Resource
+	public void setCommentsDao(CommentsDao commentsDao) {
+		this.commentsDao = commentsDao;
+	}
+	
+	@Resource
+	public void setArticleDao(ArticleDao articleDao) {
+		this.articleDao = articleDao;
+	}
+
 	public UserDao getUserDao() {
 		return userDao;
 	}
+	
 	@Resource
 	public void setUserDao(UserDao userDao) {
 		this.userDao = userDao;
 	}
 
 	@Override
-	public List<User> query(User user) {
-		return userDao.query(user);
+	public List<Map<String, Object>> query() {
+		return userDao.query();
 	}
 
 	@Override
 	public User exist(User user) {
-		List<User> userList = userDao.query(user);
-		if(userList == null){
-			return null;
-		}
-		return userList.get(0);
+		return userDao.queryExist(user);
 	}
-
+	@Override
+	public User queryById(Integer id) {
+		Map<String, Object> userMap = userDao.queryById(id);
+		User user = new User();
+		user.setId((Integer)userMap.get("id"));
+		user.setName((String)userMap.get("name"));
+		user.setEmail((String)userMap.get("email"));
+		user.setLinkURL((String)userMap.get("linkURL"));
+		user.setIdentity((Integer)userMap.get("identity"));
+		return user;
+	}
 	@Override
 	public boolean update(User user) {
 		return userDao.update(user);
@@ -42,7 +67,18 @@ public class UserManagerImpl implements UserManager{
 
 	@Override
 	public boolean delete(User user) {
-		return userDao.delete(user);
+		Comments comments = commentsDao.queryByUserId(user.getId());
+		if(comments != null)
+			commentsDao.delete(comments);
+		System.out.println("comments delete");
+		Article article = articleDao.queryByUserId(user.getId());
+		if(article != null)
+			articleDao.delete(article);
+		System.out.println("article delete");
+		if(userDao.queryById(user.getId()) != null)
+			userDao.delete(user);
+		System.out.println("delete all");
+		return true;
 	}
 
 	@Override
